+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## unreleased
+
+Features:
+
+- Default values are now validated through marshmallow field deserialization.
+  This ensures the return type is always correct ([#423](https://github.com/sloria/environs/issues/423)).
+  Thanks [lucas-bremond](https://github.com/lucas-bremond) for reporting.
+
+Other changes:
+
+- Drop support for marshmallow 3. marshmallow>=4.0.0 is now required.
+
 ## 14.6.0 (2026-02-19)
 
 Bug fixes: